READ.ME
The goal here is to make more transparent the budget data from portuguese speaking countries from Africa (Angola, Bissau Guinea, Mozambique, Cape Verde and Sao Tome e Principe) and the East Timor - the PALOP-TL.  
The budget information from those countries are published mainly in pdf tables available at the official national reports. It is hard to work and analise the data in such format. So I am building here a code that transform these pdf table into a structured data base, so we can use. 
It is important to note that small changes in the table format can change the result. So, every time you use it fora new report, investigate for changes, and check if the code needs some adjusts. Contributions are welcome! Remember to register the year and report and table the code was adapted for.